Separate out options and consolidate duplicate code
