+++ src/toil/options/wdl.py
@@ -1,0 +1,28 @@
+from argparse import ArgumentParser
+
+from configargparse import SUPPRESS
+
+
+def add_wdl_options(parser: ArgumentParser, suppress: bool = True) -> None:
+    """
+    Add WDL options to a parser. This only adds nonpositional WDL arguments
+    :param parser: Parser to add options to
+    :param suppress: Suppress help output
+    :return: None
+    """
+    suppress_help = SUPPRESS if suppress else None
+    # include arg names without a wdl specifier if suppress is False, whi
+    output_dialect_arguments = ["--wdlOutputDialect"] + (["--outputDialect"] if not suppress else [])
+    parser.add_argument(*output_dialect_arguments, dest="output_dialect", type=str, default='cromwell',
+                        choices=['cromwell', 'miniwdl'],
+                        help=suppress_help or ("JSON output format dialect. 'cromwell' just returns the workflow's "
+                                               "output values as JSON, while 'miniwdl' nests that under an 'outputs' "
+                                               "key, and includes a 'dir' key where files are written."))
+    output_directory_arguments = ["--wdlOutputDirectory"] + (["--outputDirectory", "-o"] if not suppress else [])
+    parser.add_argument(*output_directory_arguments, dest="output_directory", type=str, default=None,
+                        help=suppress_help or (
+                            "Directory or URI prefix to save output files at. By default a new directory is created "
+                            "in the current directory."))
+    output_file_arguments = ["--wdlOutputFile"] + (["--outputFile", "-m"] if not suppress else [])
+    parser.add_argument(*output_file_arguments, dest="output_file", type=str, default=None,
+                        help=suppress_help or "File or URI to save output JSON to.")
