+++ src/toil/provisioners/clusterScaler.py
@@ -32,7 +32,8 @@ from toil.batchSystems.abstractBatchSystem import (AbstractBatchSystem,
                                                    AbstractScalableBatchSystem,
                                                    NodeInfo)
 from toil.bus import ClusterDesiredSizeMessage, ClusterSizeMessage
-from toil.common import Config, defaultTargetTime
+from toil.common import Config
+from toil.options.common import defaultTargetTime
 from toil.job import JobDescription, ServiceJobDescription
 from toil.lib.conversions import bytes2human, human2bytes
 from toil.lib.retry import old_retry
